post to /archive can now handle unarchiving

+++ app/services/bucket_service.rb
@@ -1,11 +1,15 @@
 class BucketService
   def self.archive(bucket:, exclude_author_from_email_notifications: false)
-    bucket.update(archived_at: DateTime.now.utc)
-    if bucket.status == 'live'
-      bucket.contributors(exclude_author: exclude_author_from_email_notifications).each do |funder|
-        UserMailer.notify_funder_that_bucket_was_archived(funder: funder, bucket: bucket).deliver_now
+    if !bucket.archived_at
+      bucket.update(archived_at: DateTime.now.utc)
+      if bucket.status == 'live'
+        bucket.contributors(exclude_author: exclude_author_from_email_notifications).each do |funder|
+          UserMailer.notify_funder_that_bucket_was_archived(funder: funder, bucket: bucket).deliver_now
+        end
+        bucket.contributions.destroy_all
       end
-      bucket.contributions.destroy_all
+    else
+      bucket.update(archived_at: nil)
     end
   end
 end
